# Improve caching performance by adjusting default call pattern