--- a/chandra_aca/star_probs.py
+++ b/chandra_aca/star_probs.py
@@ -555,8 +555,7 @@ def grid_model_acq_prob(
 
     """
     # Get the grid model function and model parameters from a FITS file. This function
-    # call is cached. If ``model is None``` then don't pass any args so that caching
-    # matches the other in this module of ``get_grid_func_model()``.
+    # call is cached.
     gfm = get_grid_func_model(model)
 
     func_no_1p5 = gfm["func_no_1p5"]