## Description

In default usage @jeanconn noted that `get_grid_func_model` was being called once with `get_grid_func_model()` and once with `get_grid_func_model(None)`, thus requiring two reads of the model file. This fixes that.

## Interface impacts
<!-- API changes, file format updates, coordination of changes with the community. -->
None

## Testing
<!-- If relevant describe any special setup for testing. -->

### Unit tests
<!-- At least one of these must be checked if unit tests exist. DELETE the unchecked/untested options. -->
- [x] Mac

Independent check of unit tests by [REVIEWER NAME]
- [ ] [PLATFORM]:

### Functional tests
<!-- Describe and document results of any functional tests, otherwise leave the text below -->
```
In [1]: from chandra_aca import star_probs

In [2]: %time model = star_probs.get_grid_func_model()
CPU times: user 23.5 ms, sys: 24.6 ms, total: 48.2 ms
Wall time: 82.8 ms

In [3]: %time model = star_probs.get_grid_func_model(None)
CPU times: user 27 µs, sys: 5 µs, total: 32 µs
Wall time: 34.8 µs

In [4]: %time model = star_probs.get_grid_func_model("grid-*")
CPU times: user 20.4 ms, sys: 18.4 ms, total: 38.8 ms
Wall time: 68.2 ms
```