@javierggt 2023-08-18
This solves this specific case, but a way to fix it in general would be to do it in [chandra_aca/star_probs](https://github.com/sot/chandra_aca/blob/f09c50c4c7e7a661efd24af3de998592ac9dc5da/chandra_aca/star_probs.py#L422).

One could add a private function called `get_grid_func_model_` which is cached, and make `get_grid_func_model` a thin wrapper around it that is not cached. The wrapper function does what you did here and passes the arguments to `get_grid_func_model_`.