@jeanconn 2023-08-18
Which other cases are you thinking about @javierggt ?  Because we're in chandra_aca/star_probs.  I was thinking that another way to go would be to update the caching function to also do things like make func(), func(defaultmodel), and func(model=defaultmodel) all equivalent in the cache but figured that this fix was good enough.

I also figured it was basically in the noise for run speed but do figure for the use case of multiprocessing that extra reads in the job become a slightly more significant part of run time (than the FOT current use case of doing the star catalogs sequentially in one process).