@jeanconn 2023-08-18
Though I haven't checked that this fix works.  By inspection I'm not sure how it would.  I had been thinking the alternate of `gfm = get_grid_func_model(info[“default_model”])` for the default model info lookup if that is all it is supposed to do anyway.