@javierggt 2023-08-18
I wasn't thinking of any other cases. My comment was a general one. Basically, we have one function, and the calling code is doing something based on its internals, which in my opinion is not nice.

The change I propose would move the one line to massage the input into the function and would not be much more complex.

Clearly, the next step is to consider other ways in which the input arguments can be redundant, as you say, but I was not proposing that. I was proposing just to put this change in a better place, so it would not be more complex.

It's just a suggestion.