@taldcroft 2023-08-18
I like @javierggt 's idea and it is actually quite simple. Stand by.