@taldcroft 2023-08-18
BTW this is a great candidate for the `Squash and merge` button. Four meandering commits for about as many code line changes. We should start squashing PR's that have excessive commits.