@javierggt 2023-08-18
I'm not sure this fixes the problem anymore. It seems to me that the original fix was lost in the latest commit (there is not check for redundancy in the parameters). I would have expected a check in `get_grid_func_model` before calling `get_grid_func_model_`. Did you test again? I'm going to test now.