@taldcroft 2023-08-18
`get_grid_func_model()` and `get_grid_func_model(None)` both call `_get_grid_func_model(None, None, None)`.