@javierggt 2023-08-18
right, I missed that. The argument is filled by `None` by default, of course. What I just mentioned were checks we would put if we wanted something fancier, like what Jean mentioned above.