@taldcroft 2023-08-18
@javierggt - I think this is an excellent pattern you suggested for ensuring that function caching deals with potential redundancy due to keyword arg defaults and/or providing args vs kwargs. The first function call normalizes everything to a fixed list of args.