@taldcroft 2023-08-18
In your pytest output there are three different folders (CHANDRA_MODELS_REPO_DIR), so those should NOT be cached. I think that is a success.