@jeanconn 2023-08-18
But why would it ready the probability model again?  get_aca_catalog is only called once.