@jeanconn 2023-08-18
I still thought the fundamental issue was that at https://github.com/sot/chandra_aca/blob/bd9bafb9abeafa34657821d6defb5a1b36308a2b/chandra_aca/star_probs.py#L559 - model is not None.