@jeanconn 2023-08-18
I don't understand how your answer applies for the case when model isn't None.