@javierggt 2023-08-18
I guess you are focusing on the third call, with `model="grid-*"`, right? I think this solution does not fix that case, and neither did the previous solution.