@jeanconn 2023-08-18
My issue is that I'm still seeing for the completely standard use case that the model is read twice
```
In [1]: from proseco import get_aca_catalog

In [2]: get_aca_catalog(23456)
file read from /Users/jean/ska/data/chandra_models/chandra_models/xija/aca/aca_spec.json
file read from /Users/jean/ska/data/chandra_models/chandra_models/aca_acq_prob
file read from /Users/jean/ska/data/chandra_models/chandra_models/aca_acq_prob
Out[2]: 
<ACATable length=11>
 slot  idx      id    type  sz   p_acq    mag    maxmag   yang     zang    dim   res  halfw
int64 int64   int64   str3 str3 float64 float64 float64 float64  float64  int64 int64 int64
----- ----- --------- ---- ---- ------- ------- ------- -------- -------- ----- ----- -----
    0     1         2  FID  8x8   0.000    7.00    8.00  -773.20 -1742.03     1     1    25
    1     2         4  FID  8x8   0.000    7.00    8.00  2140.23   166.63     1     1    25
    2     3         5  FID  8x8   0.000    7.00    8.00 -1826.28   160.17     1     1    25
    3     4 670829768  BOT  8x8   0.977    7.51    9.01  1689.22  1326.24    28     1   160
    4     5 670317320  BOT  8x8   0.976    7.94    9.44  1988.29   539.09    28     1   160
    5     6 670312032  BOT  8x8   0.973    8.13    9.63 -1066.82 -1513.73    28     1   160
    6     7 670302264  BOT  8x8   0.971    8.21    9.71 -2277.15  1868.08    28     1   160
    7     8 670307416  BOT  8x8   0.939    8.95   10.45 -2138.51  1195.58    28     1   160
    0     9 670317248  ACQ  8x8   0.928    9.08   10.58 -1090.23 -1853.70    28     1   160
    1    10 670302816  ACQ  8x8   0.931    9.19   10.69 -1946.22  1927.59    20     1   120
    2    11 670318128  ACQ  8x8   0.756    9.84   11.09   250.49   632.72    28     1   160
```
That's what I suggested to fix on slack and thought that this PR was trying to solve.