@taldcroft 2023-08-18
> file read from /Users/jean/ska/data/chandra_models/chandra_models/aca_acq_prob
> file read from /Users/jean/ska/data/chandra_models/chandra_models/aca_acq_prob

Humm. Your pytest result is completely expected, but this one surprises me.