@jeanconn 2023-08-18
I don't disagree (with myself!) that it would also be great to normalize these cache args is also great, but I thought https://github.com/sot/chandra_aca/blob/bd9bafb9abeafa34657821d6defb5a1b36308a2b/chandra_aca/star_probs.py#L281 made this awkward.