@jeanconn 2023-08-18
I also don't see where the pytest result is completely expected since the acq probs were only "asked for" by the get_aca_catalog once.  I just figured they cloned again because of the different args (that should not be different) but I'll need to look at that in more detail.