@javierggt 2023-08-18
so... the first time, it is called with `model="grid-*"` and the second time with `model=None`.

The question is whether at this level (within `chandra_aca.star_probs`) one can know that these two calls will be the same, and whether that can change during the run (e.g. by changing an entry in `os.environ`)