@jeanconn 2023-08-18
My point was that if at https://github.com/sot/chandra_aca/blob/bd9bafb9abeafa34657821d6defb5a1b36308a2b/chandra_aca/star_probs.py#L359 you are already trying to get the default, that the pass thru to get_grid_func_model could just be info["default_model"] .  It feels consistent with what the function is trying to do anyway.