@javierggt 2023-08-18
My first inclination is to move the check
```
if model is None: 
    model = conf.default_model
```
into the wrapping function