@jeanconn 2023-08-18
Yeah, something like that probably works better to get the labels that I think were the point of the assignment.